// src/app/shared/models/createevent.ts
export class Createevent {
  event_status_code: number;
  event_type_code: number;
  commercial_type: string;
  organizer_id: number;
  venue_id: number;
  venue_name: string;
  event_name: number;
  event_start_time: string;
  event_end_time: string;
  timezone: string;
  number_of_participants: number;
  derived_days_duration: number;
  event_cost: number;
  discount: number;
  event_photo_url: string;
  comments: string;
}